Create a reusable log context listener to be used for slf4j, log4j1 and log4j2
