+++ dd-java-agent/instrumentation/log4j1/src/main/java/datadog/trace/instrumentation/log4j1/Log4j1MDCInstrumentation.java
@@ -6,13 +6,11 @@ import static net.bytebuddy.matcher.ElementMatchers.named;
 
 import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
+import datadog.trace.agent.tooling.log.LogContextScopeListener;
 import datadog.trace.api.Config;
-import datadog.trace.api.CorrelationIdentifier;
 import datadog.trace.api.GlobalTracer;
-import datadog.trace.context.ScopeListener;
 import java.lang.reflect.Method;
 import java.util.Map;
-import lombok.extern.slf4j.Slf4j;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
@@ -46,7 +44,7 @@ public class Log4j1MDCInstrumentation extends Instrumenter.Default {
 
   @Override
   public String[] helperClassNames() {
-    return new String[] {MDCContextAdvice.class.getName() + "$MDCScopeListener"};
+    return new String[] {LogContextScopeListener.class.getName()};
   }
 
   public static class MDCContextAdvice {
@@ -60,42 +58,8 @@ public class Log4j1MDCInstrumentation extends Instrumenter.Default {
         Class<?> mdcClass = instance.getClass();
         final Method putMethod = mdcClass.getMethod("put", String.class, Object.class);
         final Method removeMethod = mdcClass.getMethod("remove", String.class);
-        GlobalTracer.get().addScopeListener(new MDCScopeListener(putMethod, removeMethod));
+        GlobalTracer.get().addScopeListener(new LogContextScopeListener(putMethod, removeMethod));
       } catch (final NoSuchMethodException e) {
-      }
-    }
-
-    @Slf4j
-    public static class MDCScopeListener implements ScopeListener {
-      private final Method putMethod;
-      private final Method removeMethod;
-
-      public MDCScopeListener(final Method putMethod, final Method removeMethod) {
-        System.out.println("Initializing scope listener");
-        this.putMethod = putMethod;
-        this.removeMethod = removeMethod;
-      }
-
-      @Override
-      public void afterScopeActivated() {
-        try {
-          putMethod.invoke(
-              null, CorrelationIdentifier.getTraceIdKey(), CorrelationIdentifier.getTraceId());
-          putMethod.invoke(
-              null, CorrelationIdentifier.getSpanIdKey(), CorrelationIdentifier.getSpanId());
-        } catch (final Exception e) {
-          log.debug("Exception setting thread context context", e);
-        }
-      }
-
-      @Override
-      public void afterScopeClosed() {
-        try {
-          removeMethod.invoke(null, CorrelationIdentifier.getTraceIdKey());
-          removeMethod.invoke(null, CorrelationIdentifier.getSpanIdKey());
-        } catch (final Exception e) {
-          log.debug("Exception removing thread context context", e);
-        }
       }
     }
   }
