+++ dd-java-agent/instrumentation/log4j2/src/main/java/datadog/trace/instrumentation/log4j2/ThreadContextInstrumentation.java
@@ -6,14 +6,12 @@ import static net.bytebuddy.matcher.ElementMatchers.named;
 
 import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
+import datadog.trace.agent.tooling.log.LogContextScopeListener;
 import datadog.trace.api.Config;
-import datadog.trace.api.CorrelationIdentifier;
 import datadog.trace.api.GlobalTracer;
-import datadog.trace.context.ScopeListener;
 import java.lang.reflect.Method;
 import java.security.ProtectionDomain;
 import java.util.Map;
-import lombok.extern.slf4j.Slf4j;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
@@ -60,7 +58,7 @@ public class ThreadContextInstrumentation extends Instrumenter.Default {
 
   @Override
   public String[] helperClassNames() {
-    return new String[] {ThreadContextAdvice.class.getName() + "$ThreadContextScopeListener"};
+    return new String[] {LogContextScopeListener.class.getName()};
   }
 
   public static class ThreadContextAdvice {
@@ -69,44 +67,10 @@ public class ThreadContextInstrumentation extends Instrumenter.Default {
       try {
         final Method putMethod = threadClass.getMethod("put", String.class, String.class);
         final Method removeMethod = threadClass.getMethod("remove", String.class);
-        GlobalTracer.get()
-            .addScopeListener(new ThreadContextScopeListener(putMethod, removeMethod));
+        GlobalTracer.get().addScopeListener(new LogContextScopeListener(putMethod, removeMethod));
       } catch (final NoSuchMethodException e) {
         org.slf4j.LoggerFactory.getLogger(threadClass)
             .debug("Failed to add log4j ThreadContext span listener", e);
-      }
-    }
-
-    @Slf4j
-    public static class ThreadContextScopeListener implements ScopeListener {
-      private final Method putMethod;
-      private final Method removeMethod;
-
-      public ThreadContextScopeListener(final Method putMethod, final Method removeMethod) {
-        this.putMethod = putMethod;
-        this.removeMethod = removeMethod;
-      }
-
-      @Override
-      public void afterScopeActivated() {
-        try {
-          putMethod.invoke(
-              null, CorrelationIdentifier.getTraceIdKey(), CorrelationIdentifier.getTraceId());
-          putMethod.invoke(
-              null, CorrelationIdentifier.getSpanIdKey(), CorrelationIdentifier.getSpanId());
-        } catch (final Exception e) {
-          log.debug("Exception setting thread context context", e);
-        }
-      }
-
-      @Override
-      public void afterScopeClosed() {
-        try {
-          removeMethod.invoke(null, CorrelationIdentifier.getTraceIdKey());
-          removeMethod.invoke(null, CorrelationIdentifier.getSpanIdKey());
-        } catch (final Exception e) {
-          log.debug("Exception removing thread context context", e);
-        }
       }
     }
   }
