+++ dd-java-agent/instrumentation/slf4j-mdc/src/main/java/datadog/trace/instrumentation/slf4j/mdc/MDCInjectionInstrumentation.java
@@ -6,14 +6,12 @@ import static net.bytebuddy.matcher.ElementMatchers.named;
 
 import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
+import datadog.trace.agent.tooling.log.LogContextScopeListener;
 import datadog.trace.api.Config;
-import datadog.trace.api.CorrelationIdentifier;
 import datadog.trace.api.GlobalTracer;
-import datadog.trace.context.ScopeListener;
 import java.lang.reflect.Method;
 import java.security.ProtectionDomain;
 import java.util.Map;
-import lombok.extern.slf4j.Slf4j;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.method.MethodDescription;
 import net.bytebuddy.description.type.TypeDescription;
@@ -61,7 +59,7 @@ public class MDCInjectionInstrumentation extends Instrumenter.Default {
 
   @Override
   public String[] helperClassNames() {
-    return new String[] {MDCAdvice.class.getName() + "$MDCScopeListener"};
+    return new String[] {LogContextScopeListener.class.getName()};
   }
 
   public static class MDCAdvice {
@@ -70,42 +68,9 @@ public class MDCInjectionInstrumentation extends Instrumenter.Default {
       try {
         final Method putMethod = mdcClass.getMethod("put", String.class, String.class);
         final Method removeMethod = mdcClass.getMethod("remove", String.class);
-        GlobalTracer.get().addScopeListener(new MDCScopeListener(putMethod, removeMethod));
+        GlobalTracer.get().addScopeListener(new LogContextScopeListener(putMethod, removeMethod));
       } catch (final NoSuchMethodException e) {
         org.slf4j.LoggerFactory.getLogger(mdcClass).debug("Failed to add MDC span listener", e);
-      }
-    }
-
-    @Slf4j
-    public static class MDCScopeListener implements ScopeListener {
-      private final Method putMethod;
-      private final Method removeMethod;
-
-      public MDCScopeListener(final Method putMethod, final Method removeMethod) {
-        this.putMethod = putMethod;
-        this.removeMethod = removeMethod;
-      }
-
-      @Override
-      public void afterScopeActivated() {
-        try {
-          putMethod.invoke(
-              null, CorrelationIdentifier.getTraceIdKey(), CorrelationIdentifier.getTraceId());
-          putMethod.invoke(
-              null, CorrelationIdentifier.getSpanIdKey(), CorrelationIdentifier.getSpanId());
-        } catch (final Exception e) {
-          log.debug("Exception setting mdc context", e);
-        }
-      }
-
-      @Override
-      public void afterScopeClosed() {
-        try {
-          removeMethod.invoke(null, CorrelationIdentifier.getTraceIdKey());
-          removeMethod.invoke(null, CorrelationIdentifier.getSpanIdKey());
-        } catch (final Exception e) {
-          log.debug("Exception removing mdc context", e);
-        }
       }
     }
   }
